Overhaul of DeerLab's back end solvers
Changes introduced: 

- Rename `fitregmodel` to `rlls` (regularized linear least squares).
- Rename `fitparamodel` to `nlls` (non-linear least squares).
- Merge `snlls.py`, `nlls.py`, `rnlls.py`, `lsqcomponents.py`, and `nnls.py` into a single module `solvers.py`.
- Expanded the docs of `snlls.py`, `nlls.py`, and `rnlls.py`.
- The main solvers `snlls.py`, `nlls.py`, `rnlls.py` now share the same utility functions and basic processes. This fixes bugs and strange behaviors of specific functions under certain situations.
- Deprecate the use of TV and Huber regularization, accordingly the keywords `regtype`, `huberparameter` have been removed throughout.
- Generalize the regularization operator. Back end functions no longer take `regorder` (regularization operator order) but instead `regop` (regularization operator).
- The function `rnlls.py` is now a general regularized linear least-squares solver, accepting arbitrary bounds and adapting the linear LSQ solver back end accordingly as in `snlls`.
- The function `rnlls.py` now automatically decides whether regularization is needed as in `snlls`.
- Add option `extrapenalty` to `nlls` to accept arbitrary penalties as in `snlls`.
- Add test to control that all variables in the main functions are defined in the docstring. Fixed errors in docstrings throughout detected by new tests.
- Implement a new function to ensure that estimated Jacobians are positive semi-definite matrices. This fixes the appearance of warnings and bugs when calculating confidence intervals.
