+++ deerlab/__init__.py
@@ -4,21 +4,17 @@ from .bg_models import *
 from .ex_models import *
 from .deerload import deerload
 from .selregparam import selregparam
-from .nnls import fnnls,cvxnnls
 from .regparamrange import regparamrange
 from .dipolarkernel import dipolarkernel
 from .dipolarbackground import dipolarbackground
-from .fitregmodel import fitregmodel
+from .solvers import rlls, nlls, snlls
 from .regoperator import regoperator
 from .correctphase import correctphase
 from .mixmodels import mixmodels
 from .correctzerotime import correctzerotime
 from .noiselevel import noiselevel
 from .whitegaussnoise import whitegaussnoise
-from .lsqcomponents import lsqcomponents
-from .snlls import snlls
 from .fitmultimodel import fitmultimodel
-from .fitparamodel import fitparamodel
 from .bootan import bootan
 from .fitmodel import fitmodel
 from .correctscale import correctscale
