--- docsrc/source/functions/fitparamodel.rst
@@ -1,8 +1,0 @@
-.. highlight:: python
-.. _fitparamodel:
-
-*********************
-:mod:`fitparamodel`
-*********************
-
-.. autofunction:: deerlab.fitparamodel
