--- docsrc/source/functions/fitregmodel.rst
@@ -1,10 +1,0 @@
-.. highlight:: python
-.. _fitregmodel:
-
-*********************
-:mod:`fitregmodel`
-*********************
-
-.. autofunction:: deerlab.fitregmodel
-
-
