+++ docsrc/source/functions/nlls.rst
@@ -1,0 +1,8 @@
+.. highlight:: python
+.. _nlls:
+
+*********************
+:mod:`nlls`
+*********************
+
+.. autofunction:: deerlab.nlls
