+++ docsrc/source/functions/rlls.rst
@@ -1,0 +1,10 @@
+.. highlight:: python
+.. _rlls:
+
+*********************
+:mod:`rlls`
+*********************
+
+.. autofunction:: deerlab.rlls
+
+
