+++ test/test_rlls.py
@@ -1,10 +1,11 @@
 
 import numpy as np
-from deerlab import fitregmodel,dipolarkernel, regoperator, whitegaussnoise
+from deerlab import rlls,dipolarkernel, regoperator, whitegaussnoise
 from deerlab.dd_models import dd_gauss,dd_gauss2
-from deerlab.utils import ovl, skip_on
+from deerlab.utils import ovl, skip_on, assert_docstring
 
-def assert_solver(regtype,solver):
+
+def assert_solver(solver):
 #============================================================
 
     np.random.seed(1)
@@ -13,78 +14,35 @@ def assert_solver(regtype,solver):
     P = dd_gauss(r,[3,0.2])
     K = dipolarkernel(t,r)
     V = K@P + whitegaussnoise(t,0.01)
-    fit = fitregmodel(V,K,r,regtype,'aic',solver=solver)
+    fit = rlls(V,K,lb=np.zeros_like(r),solver=solver,uq=False)
 
-    assert ovl(P,fit.P) > 0.95 # more than 95% overlap
+    assert ovl(P,fit.param) > 0.95 # more than 95% overlap
 #============================================================
 
 def test_tikh_solver_fnnls():
 #============================================================
     "Check that 'fnnls' solves a Tikhonov regularization problem"
 
-    assert_solver('tikhonov','fnnls')
+    assert_solver('fnnls')
 #============================================================
 
 def test_tikh_solver_nnlsbpp():
 #============================================================
     "Check that 'nnlsbpp' solves a Tikhonov regularization problem"
 
-    assert_solver('tikhonov','nnlsbpp')
+    assert_solver('nnlsbpp')
 #============================================================
 
 def test_tikh_solver_cvx():
 #============================================================
     "Check that 'cvx' solves a Tikhonov regularization problem"
 
-    assert_solver('tikhonov','cvx')
-#============================================================
-
-def test_tv_solver_fnnls():
-#============================================================
-    "Check that 'fnnls' solves a TV regularization problem"
-
-    assert_solver('tv','fnnls')
-#============================================================
-
-def test_tv_solver_nnlsbpp():
-#============================================================
-    "Check that 'nnlsbpp' solves a TV regularization problem"
-
-    assert_solver('tv','nnlsbpp')
-#============================================================
-
-def test_tv_solver_cvx():
-#============================================================
-    "Check that 'cvx' solves a TV regularization problem"
-
-    assert_solver('tv','cvx')
-#============================================================
-
-
-def test_huber_solver_fnnls():
-#============================================================
-    "Check that 'fnnls' solves a Huber regularization problem"
-
-    assert_solver('huber','fnnls')
-#============================================================
-
-def test_huber_solver_nnlsbpp():
-#============================================================
-    "Check that 'nnlsbpp' solves a Huber regularization problem"
-
-    assert_solver('huber','nnlsbpp')
-#============================================================
-
-def test_huber_solver_cvx():
-#============================================================
-    "Check that 'cvx' solves a Huber regularization problem"
-
-    assert_solver('huber','cvx')
+    assert_solver('cvx')
 #============================================================
 
 def test_tikh_with_noise():
 #============================================================
-    "Check the TV regularization method"
+    "Check the Tikhonov regularization method with noise"
 
     np.random.seed(1)
     t = np.linspace(0,3,200)
@@ -93,59 +51,25 @@ def test_tikh_with_noise():
     K = dipolarkernel(t,r)
     V = K@P + whitegaussnoise(t,0.01)
 
-    fit = fitregmodel(V,K,r,'tikhonov','aic')
+    fit = rlls(V,K,lb=np.zeros_like(r))
     
-    assert ovl(P,fit.P) > 0.95 # more than 95% overlap
+    assert ovl(P,fit.param) > 0.95 # more than 95% overlap
 #============================================================
-
-def test_tv_with_noise():
-#============================================================
-    "Check the TV regularization method"
-
-    np.random.seed(1)
-    t = np.linspace(0,3,200)
-    r = np.linspace(1,5,100)
-    P = dd_gauss(r,[3,0.08])
-    K = dipolarkernel(t,r)
-    V = K@P + whitegaussnoise(t,0.01)
-
-    fit = fitregmodel(V,K,r,'tv','aic')
-    
-    assert ovl(P,fit.P) > 0.95 # more than 95% overlap
-#============================================================
-
-def test_huber_with_noise():
-#============================================================
-    "Check the Huber regularization method"
-
-    np.random.seed(1)
-    t = np.linspace(0,3,200)
-    r = np.linspace(1,5,100)
-    P = dd_gauss(r,[3,0.08])
-    K = dipolarkernel(t,r)
-    V = K@P + whitegaussnoise(t,0.01)
-
-    fit = fitregmodel(V,K,r,'huber','aic')
-    
-    assert ovl(P,fit.P) > 0.95 # more than 95% overlap
-#============================================================
-
 
 def test_unconstrained():
 #============================================================
     "Check that unconstrained distributions are correctly fitted"
 
     t = np.linspace(-2,4,300)
-    r = np.linspace(2,6,100)
+    r = np.linspace(2,6,200)
     P = dd_gauss(r,[3,0.2])
     K = dipolarkernel(t,r)
-    L = regoperator(r,2)
+    L = regoperator(np.arange(len(r)),2)
     V = K@P
 
-    fit = fitregmodel(V,K,r,'tikhonov',regparam=0.05,nonnegativity=False,renormalize=False)
-    Pfit_ref = np.linalg.solve(K.T@K + 0.05**2*L.T@L, K.T@V)
-
-    assert ovl(Pfit_ref,fit.P) > 0.99 # more than 99% overlap
+    fit = rlls(V,K,regparam=0.2)
+    Pfit_ref = np.linalg.solve(K.T@K + 0.2**2*L.T@L, K.T@V)
+    assert ovl(Pfit_ref,fit.param) > 0.99 # more than 99% overlap
 #============================================================
 
 def generate_global_dataset():
@@ -194,40 +118,9 @@ def test_confinter_tikh():
     K = dipolarkernel(t,r)
     V = K@P
 
-    fit = fitregmodel(V,K,r,'tikhonov','aic')
+    fit = rlls(V,K,lb=np.zeros_like(r))
 
-    assert_confidence_intervals(fit.Puncert,fit.P)
-#============================================================
-
-def test_confinter_tv():
-#============================================================
-    "Check that the confidence intervals are correctly computed using TV regularization"
-
-    t = np.linspace(-2,4,300)
-    r = np.linspace(2,6,100)
-    P = dd_gauss(r,[3,0.2])
-    K = dipolarkernel(t,r)
-    V = K@P
-
-    fit = fitregmodel(V,K,r,'tv','aic')
-
-    assert_confidence_intervals(fit.Puncert,fit.P)
-#============================================================
-
-
-def test_confinter_huber():
-#============================================================
-    "Check that the confidence intervals are correctly computed using Huber regularization"
-
-    t = np.linspace(-2,4,300)
-    r = np.linspace(2,6,100)
-    P = dd_gauss(r,[3,0.2])
-    K = dipolarkernel(t,r)
-    V = K@P
-
-    fit = fitregmodel(V,K,r,'huber','aic')
-
-    assert_confidence_intervals(fit.Puncert,fit.P)
+    assert_confidence_intervals(fit.paramUncert,fit.param)
 #============================================================
 
 def test_confinter_global():
@@ -236,9 +129,9 @@ def test_confinter_global():
 
     r,_,V1,V2,K1,K2 = generate_global_dataset()
 
-    fit = fitregmodel([V1,V2],[K1,K2],r,'tikhonov','aic')
+    fit = rlls([V1,V2],[K1,K2],lb=np.zeros_like(r))
 
-    assert_confidence_intervals(fit.Puncert,fit.P)
+    assert_confidence_intervals(fit.paramUncert,fit.param)
 #============================================================
 
 def test_confinter_Vfit():
@@ -251,25 +144,9 @@ def test_confinter_Vfit():
     K = dipolarkernel(t,r,mod=0.2)
     V = K@P + whitegaussnoise(t,0.05)
 
-    fit = fitregmodel(V,K,r,'tikhonov','aic')
+    fit = rlls(V,K,lb=np.zeros_like(r))
 
-    assert_confidence_intervals(fit.Vuncert,fit.V)
-#============================================================
-
-def test_renormalize():
-#============================================================
-    "Check that renormalization of the distribution can be correctly enabled/disabled"
-
-    t = np.linspace(-2,4,300)
-    r = np.linspace(2,6,100)
-    P = dd_gauss(r,[3,0.2])
-    K = dipolarkernel(t,r)
-    V = K@P
-
-    fitraw = fitregmodel(V,K,r,'tikhonov','aic',renormalize = False)
-    fitrenorm = fitregmodel(V,K,r,'tikhonov','aic',renormalize = True)
-
-    assert max(abs(fitraw.P - fitrenorm.P*np.trapz(fitraw.P,r))) < 1e-8
+    assert_confidence_intervals(fit.modelUncert,fit.model)
 #============================================================
 
 def test_scale_agnostic():
@@ -283,25 +160,9 @@ def test_scale_agnostic():
     scale = 1e8
     V = scale*K@P
 
-    fit = fitregmodel(V,K,r,'tikhonov','aic',renormalize = False)
+    fit = rlls(V,K,lb=np.zeros_like(r))
 
-    assert max(abs(P - fit.P/fit.scale)) < 1e-3
-#============================================================
-
-def test_scale_fit():
-#============================================================
-    "Check agnosticity of the w.r.t. an arbitrary scale of dipolar signal"
-
-    t = np.linspace(-2,4,300)
-    r = np.linspace(2,6,100)
-    P = dd_gauss(r,[3,0.2])
-    K = dipolarkernel(t,r)
-    scale = 1e8
-    V = scale*K@P
-
-    fit = fitregmodel(V,K,r,'tikhonov','aic',renormalize = False)
-
-    assert abs(1 - scale/fit.scale) < 1e-4
+    assert max(abs(V - fit.model)/scale) < 1e-3
 #============================================================
 
 def test_nonuniform_r():
@@ -314,27 +175,9 @@ def test_nonuniform_r():
     K = dipolarkernel(t,r)
     V = K@P
 
-    fit = fitregmodel(V,K,r,'tikhonov','aic')
-    Vfit = K@fit.P
+    fit = rlls(V,K,lb=np.zeros_like(r))
+    Vfit = K@fit.param
     assert abs(Vfit[0] - 1) < 1e-6
-#============================================================
-
-
-def test_obir():
-#============================================================
-    "Check that the OBIR algorithm runs and returns a correct result"
-
-    np.random.seed(1)
-    t = np.linspace(0,3,200)
-    r = np.linspace(2,7,100)
-    P = dd_gauss2(r,[3.7,0.2,0.5,4.3,0.1,0.5])
-    K = dipolarkernel(t,r)
-
-    V = K@P + whitegaussnoise(t,0.04)
-
-    fit = fitregmodel(V,K,r,'tikhonov','aic',obir=True,noiselevelaim=0.042)
-    
-    assert ovl(P,fit.P) > 0.7 # more than 80% overlap
 #============================================================
 
 
@@ -350,7 +193,7 @@ def test_goodness_of_fit():
     sigma = 0.03
     V = K@P + whitegaussnoise(t,sigma,seed=1,rescale=True)
 
-    fit = fitregmodel(V,K,r,'tikhonov','aic',noiselvl=sigma)
+    fit = rlls(V,K,lb=np.zeros_like(r),noiselvl=sigma)
     stats = fit.stats
 
     assert abs(stats['chi2red'] - 1) < 0.05
@@ -371,7 +214,7 @@ def test_goodness_of_fit_scaled():
     sigma = V0*0.03
     V += whitegaussnoise(t,sigma,seed=1,rescale=True)
 
-    fit = fitregmodel(V,K,r,'tikhonov','aic',noiselvl=sigma)
+    fit = rlls(V,K,lb=np.zeros_like(r),noiselvl=sigma)
     stats = fit.stats
 
     assert abs(stats['chi2red'] - 1) < 0.05
@@ -392,28 +235,10 @@ def test_goodness_of_fit_global():
     np.random.seed(2)
     V2 = K2@P + whitegaussnoise(t2,0.01)
 
-    fit = fitregmodel([V1,V2],[K1,K2],r,'tikhonov','aic')
+    fit = rlls([V1,V2],[K1,K2],lb=np.zeros_like(r))
     stats = fit.stats
 
     assert (abs(stats[0]['R2'] - 1) < 5e-2) and (abs(stats[1]['R2'] - 1) < 5e-2)
-#============================================================
-
-def test_globalfit_scales():
-#============================================================
-    "Check that the global fit with arbitrary amplitudes works."
-    t1 = np.linspace(0,5,300)
-    t2 = np.linspace(0,2,300)
-    r = np.linspace(3,5,100)
-    P = dd_gauss(r,[4,0.25])
-    K1 = dipolarkernel(t1,r) 
-    K2 = dipolarkernel(t2,r)
-    scales = [1e3, 1e9]
-    V1 = scales[0]*K1@P
-    V2 = scales[1]*K2@P
-
-    fit = fitregmodel([V1,V2],[K1,K2],r,'tikhonov','aic',weights=[1,1],renormalize=False)
-
-    assert max(abs(np.asarray(scales)/np.asarray(fit.scale) - 1)) < 1e-2 
 #============================================================
 
 @skip_on('_tkinter.TclError', reason="A problem with the Tk backend occured")
@@ -427,7 +252,7 @@ def test_plot():
     K = dipolarkernel(t,r)
     V = K@P
 
-    fit = fitregmodel(V,K,r,'tikhonov','aic')
+    fit = rlls(V,K,lb=np.zeros_like(r))
     fig = fit.plot(show=False)
     assert str(fig.__class__)=="<class 'matplotlib.figure.Figure'>"
 #============================================================
@@ -442,7 +267,7 @@ def test_cost_value():
     P = dd_gauss(r,[3,0.2])
     K = dipolarkernel(t,r)
     V = K@P + whitegaussnoise(t,0.01)
-    fit = fitregmodel(V,K,r)
+    fit = rlls(V,K,lb=np.zeros_like(r))
 
     assert isinstance(fit.cost,float) and np.round(fit.cost/np.sum(fit.residuals**2),5)==1
 #============================================================
@@ -457,9 +282,9 @@ def test_convergence_criteria():
     K = dipolarkernel(t,r)
     V = K@P
 
-    fit = fitregmodel(V,K,r,'tikhonov','aic',tol=1e-9,maxiter=2e3)
+    fit = rlls(V,K,tol=1e-9,maxiter=2e3,lb=np.zeros_like(r))
   
-    assert ovl(P,fit.P) > 0.90 # more than 80% overlap
+    assert ovl(P,fit.param) > 0.90 # more than 80% overlap
 #============================================================
 
 def test_confinter_values():
@@ -480,10 +305,10 @@ def test_confinter_values():
                 [2.878254646104185, 3.00073387702068], 
                 [2.9089331341860465, 2.9700584546043713]]
     
-    fit = fitregmodel(y,A, np.arange(2),renormalize=False,nonnegativity=False,regparam=0,regorder=0)
-    a_ci = [fit.Puncert.ci(cov[i])[0,:] for i in range(3)]
-    b_ci = [fit.Puncert.ci(cov[i])[1,:] for i in range(3)]
-    print(np.max(abs(np.array(a_ci) - np.array(a_ci_ref)))/p[0])
+    fit = rlls(y,A,reg=False)
+    a_ci = [fit.scale*fit.paramUncert.ci(cov[i])[0,:] for i in range(3)]
+    b_ci = [fit.scale*fit.paramUncert.ci(cov[i])[1,:] for i in range(3)]
+    
     ci_match = lambda ci,ci_ref,truth:np.max(abs(np.array(ci) - np.array(ci_ref)))/truth < 0.05
     assert ci_match(a_ci,a_ci_ref,p[0]) & ci_match(b_ci,b_ci_ref,p[1])
 # ======================================================================
@@ -503,10 +328,10 @@ def test_global_weights():
     V1 = K@P1 + whitegaussnoise(t,0.01,seed=1)
     V2 = K@P2 + whitegaussnoise(t,0.01,seed=1)
 
-    fit1 = fitregmodel([V1,V2],[K,K],r,weights=[1,1e-10])
-    fit2 = fitregmodel([V1,V2],[K,K],r,weights=[1e-10,1])
+    fit1 = rlls([V1,V2],[K,K],lb=np.zeros_like(r),weights=[1,1e-10])
+    fit2 = rlls([V1,V2],[K,K],lb=np.zeros_like(r),weights=[1e-10,1])
 
-    assert ovl(P1,fit1.P) > 0.95 and ovl(P2,fit2.P) > 0.95
+    assert ovl(P1,fit1.param) > 0.95 and ovl(P2,fit2.param) > 0.95
 # ======================================================================
 
 def test_global_weights_default():
@@ -522,7 +347,14 @@ def test_global_weights_default():
     V1 = scales[0]*K@P + whitegaussnoise(t,0.001,seed=1)
     V2 = scales[1]*K@P + whitegaussnoise(t,0.1,seed=1)
     
-    fit = fitregmodel([V1,V2],[K,K],r)
+    fit = rlls([V1,V2],[K,K],lb=np.zeros_like(r))
 
-    assert ovl(P,fit.P)>0.95
+    assert ovl(P,fit.param)>0.95
+# ======================================================================
+
+
+def test_docstring():
+# ======================================================================
+    "Check that the docstring includes all variables and keywords."
+    assert_docstring(rlls)
 # ======================================================================
